+++ src/Microsoft.Extensions.Localization/Internal/POParser.cs
@@ -10,88 +10,6 @@ using Microsoft.Extensions.Localization.Internal.POLines;
 
 namespace Microsoft.Extensions.Localization.Internal
 {
-    public class POEntry
-    {
-        public StringBuilder OriginalBuilder { get; private set; }
-
-        public static StringBuilder ConcatStringBuilders(StringBuilder first, StringBuilder second)
-        {
-            for (int i = 0; i < second.Length; i++)
-            {
-                first.Append(second[i]);
-            }
-
-            return first;
-        }
-
-        public void AppendOriginalLine(StringBuilder line)
-        {
-            if (OriginalBuilder == null)
-            {
-                OriginalBuilder = new StringBuilder();
-            }
-
-            ConcatStringBuilders(OriginalBuilder, line);
-        }
-
-        private string _original;
-
-        public string Original
-        {
-            get
-            {
-                if (_original == null)
-                {
-                    _original = OriginalBuilder?.ToString();
-                }
-
-                return _original;
-            }
-        }
-
-        public string OriginalPlural { get; set; }
-
-        public StringBuilder TranslationBuilder { get; private set; }
-
-        public void AppendTranslationLine(StringBuilder line)
-        {
-            if (TranslationBuilder == null)
-            {
-                TranslationBuilder = new StringBuilder();
-            }
-
-            _translation = null;
-            ConcatStringBuilders(TranslationBuilder, line);
-        }
-
-        private string _translation;
-
-        public string Translation
-        {
-            get
-            {
-                if (_translation == null)
-                {
-                    _translation = TranslationBuilder?.ToString();
-                }
-
-                return _translation;
-            }
-        }
-
-        public IDictionary<int, string> TranslationPlurals { get; set; } = new Dictionary<int, string>();
-
-        public string Untranslated { get; set; }
-
-        public List<string> References { get; set; } = new List<string>();
-
-        public List<string> Contexts { get; set; } = new List<string>();
-
-        public List<string> Flags { get; set; } = new List<string>();
-
-        public string Comment { get; set; }
-    }
-
     public class POParser
     {
         private static IList<TokenLine> Lines = new List<TokenLine>
